+++ cocotb/triggers.py
@@ -36,6 +36,7 @@ from numbers import Real
 from typing import Optional, Union
 
 import cocotb
+import cocotb.task
 from cocotb import outcomes, simulator
 from cocotb.log import SimLog
 from cocotb.utils import (
@@ -778,7 +779,7 @@ class _AggregateWaitable(Waitable):
 
         # Do some basic type-checking up front, rather than waiting until we
         # await them.
-        allowed_types = (Trigger, Waitable, cocotb.decorators.Task)
+        allowed_types = (Trigger, Waitable, cocotb.task.Task)
         for trigger in self.triggers:
             if not isinstance(trigger, allowed_types):
                 raise TypeError(
@@ -793,7 +794,7 @@ class _AggregateWaitable(Waitable):
         return "{}({})".format(
             type(self).__qualname__,
             ", ".join(
-                repr(Join(t)) if isinstance(t, cocotb.decorators.Task) else repr(t)
+                repr(Join(t)) if isinstance(t, cocotb.task.Task) else repr(t)
                 for t in self.triggers
             ),
         )
@@ -973,7 +974,7 @@ async def with_timeout(trigger, timeout_time, timeout_unit="step"):
         await with_timeout(First(coro, event.wait()), 100, 'ns')
 
     Args:
-        trigger (:class:`~cocotb.triggers.Trigger`, :class:`~cocotb.triggers.Waitable`, :class:`~cocotb.decorators.Task`, or :term:`python:coroutine`):
+        trigger (:class:`~cocotb.triggers.Trigger`, :class:`~cocotb.triggers.Waitable`, :class:`~cocotb.task.Task`, or :term:`python:coroutine`):
             A single object that could be right of an :keyword:`await` expression in cocotb.
         timeout_time (numbers.Real or decimal.Decimal):
             Simulation time duration before timeout occurs.
