+++ documentation/source/newsfragments/2876.change.1.rst
@@ -1,1 +1,1 @@
-Renamed ``RunningTask`` to :class:`~cocotb.decorators.Task`.
+Renamed ``RunningTask`` to :class:`~cocotb.task.Task`.
