+++ documentation/source/newsfragments/2876.change.rst
@@ -1,1 +1,1 @@
-Made :class:`~cocotb.decorators.Task` interface more like :py:class:`asyncio.Task`'s.
+Made :class:`~cocotb.task.Task` interface more like :py:class:`asyncio.Task`'s.
